fix(cli): honor trajectory pointer files in sessions tail
